Division du server
